Added mail function for reporting errors. Only implementation I've seen that works uses sendmail, there are some notes in the function.
Added a command line option to set a directory for the neurospaces_cron.stdout on the command
line. 

Added a function for copying over all of the text_* files generated in /tmp
to the output directory set on the command line. Does nothing if not set. 

Added a function that removes all of the generated files from the tests (neurospaces_check)
that are put into the /tmp directory.